improve tag matching
- keep backwards compatibility so that Repos without tags return an empty set and empty set means "matches"
- renamed Tags.matchesTags to Tags.isIncluded
- move tags to own package aQute.bnd.service.tags
