+++ biz.aQute.bndlib/src/aQute/bnd/service/tags/package-info.java
@@ -1,0 +1,4 @@
+@Version("1.0.0")
+package aQute.bnd.service.tags;
+
+import org.osgi.annotation.versioning.Version;
