+++ biz.aQute.repository/src/aQute/bnd/deployer/repository/LocalIndexedRepo.java
@@ -1,7 +1,6 @@
 package aQute.bnd.deployer.repository;
 
 import static aQute.bnd.deployer.repository.RepoConstants.DEFAULT_CACHE_DIR;
-import static aQute.bnd.service.Tags.parse;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
@@ -41,6 +40,7 @@ import aQute.bnd.service.Refreshable;
 import aQute.bnd.service.RepositoryListenerPlugin;
 import aQute.bnd.service.ResourceHandle;
 import aQute.bnd.service.ResourceHandle.Location;
+import aQute.bnd.service.tags.Tags;
 import aQute.bnd.version.Version;
 import aQute.bnd.version.VersionRange;
 import aQute.lib.hex.Hex;
@@ -135,7 +135,7 @@ public class LocalIndexedRepo extends AbstractIndexedRepo implements Refreshable
 				String.format("Cannot create repository cache: '%s' already exists but is not directory.",
 					cacheDir.getAbsolutePath()));
 
-		super.setTags(parse(map.get(PROP_TAGS), DEFAULT_REPO_TAGS));
+		super.setTags(Tags.parse(map.get(PROP_TAGS), DEFAULT_REPO_TAGS));
 	}
 
 	@Override
